Extract `ExportSummary` class for account object counts
