+++ app/controllers/settings/exports_controller.rb
@@ -9,7 +9,7 @@ class Settings::ExportsController < Settings::BaseController
   skip_before_action :require_functional!
 
   def show
-    @export  = Export.new(current_account)
+    @export_summary = ExportSummary.new(preloaded_account)
     @backups = current_user.backups
   end
 
@@ -24,5 +24,16 @@ class Settings::ExportsController < Settings::BaseController
     BackupWorker.perform_async(backup.id)
 
     redirect_to settings_export_path
+  end
+
+  private
+
+  def preloaded_account
+    current_account.tap do |account|
+      ActiveRecord::Associations::Preloader.new(
+        records: [account],
+        associations: :account_stat
+      ).call
+    end
   end
 end
