+++ app/models/export.rb
@@ -55,42 +55,6 @@ class Export
     end
   end
 
-  def total_storage
-    account.media_attachments.sum(:file_file_size)
-  end
-
-  def total_statuses
-    account.statuses_count
-  end
-
-  def total_bookmarks
-    account.bookmarks.count
-  end
-
-  def total_follows
-    account.following_count
-  end
-
-  def total_lists
-    account.owned_lists.count
-  end
-
-  def total_followers
-    account.followers_count
-  end
-
-  def total_blocks
-    account.blocking.count
-  end
-
-  def total_mutes
-    account.muting.count
-  end
-
-  def total_domain_blocks
-    account.domain_blocks.count
-  end
-
   private
 
   def to_csv(accounts)
